Admin: Implemented mapping email to reg_id

+++ app/Http/Controllers/Client/RegSysClientController.php
@@ -10,34 +10,67 @@ use Illuminate\Support\Facades\Log;
 
 class RegSysClientController extends Controller
 {
-    public static function bookPackage(string $id, TableType $tableType): bool
+    public static function getPackages(string|null $reg_id): mixed
     {
+        if (!self::validateParam($reg_id, 'reg id')) {
+            return [];
+        }
+
+        $response = Http::get(config('services.regsys.url') . '/package-api', [
+            'token' => config('services.regsys.token'),
+            'id' => $reg_id,
+        ]);
+
+        if ($response->ok()) {
+            return $response->json('packages');
+        } else {
+            self::logError("Registration with id " . $reg_id . " could not be retrieved, reason: " . $response->reason());
+            return [];
+        }
+    }
+
+    public static function bookPackage(string|null $reg_id, TableType $tableType): bool
+    {
+        if (!self::validateParam($reg_id, 'reg id')) {
+            return false;
+        }
+        if (!self::validateParam($tableType, 'table type')) {
+            return false;
+        }
+
         $response = Http::post(config('services.regsys.url') . '/package-api', [
             'token' => config('services.regsys.token'),
-            'id' => $id,
+            'id' => $reg_id,
             'package' => $tableType->package,
         ]);
 
         if ($response->ok()) {
             return true;
         } else {
-            Log::warning("Package for id " . $id . " could not be booked, reason: " . $response->reason());
+            self::logError("Package for id " . $reg_id . " could not be booked, reason: " . $response->reason());
             return false;
         }
     }
 
-    public static function removePackage(string $id, TableType $tableType): bool
+    public static function removePackage(string|null $reg_id, TableType $tableType): bool
     {
+        if (!self::validateParam($reg_id, 'reg id')) {
+            return false;
+        }
+        if (!self::validateParam($tableType, 'table type')) {
+            return false;
+        }
+
         $response = Http::delete(config('services.regsys.url') . '/package-api', [
             'token' => config('services.regsys.token'),
-            'id' => $id,
+            'id' => $reg_id,
             'package' => $tableType->package,
         ]);
 
         if ($response->ok()) {
             return true;
         } else {
-            Log::warning("Package for id " . $id . " could not be deleted, reason: " . $response->reason());
+            self::logError("Package for id " . $reg_id . " could not be deleted, reason: " . $response->reason());
             return false;
         }
     }
@@ -49,25 +82,44 @@ class RegSysClientController extends Controller
         ]);
 
         if ($response->ok()) {
-            return $response->json();
+            return $response->json('result');
         } else {
-            Log::warning("Registrations could not be retrieved, reason: " . $response->reason());
+            self::logError("Registrations could not be retrieved, reason: " . $response->reason());
+            return [];
+        }
+    }
+
+    public static function getSingleReg(string|null $reg_id): mixed
+    {
+        if (!self::validateParam($reg_id, 'reg id')) {
+            return null;
+        }
+
+        $response = Http::get(config('services.regsys.url') . '/dealers-den-api', [
+            'token' => config('services.regsys.token'),
+            'id' => $reg_id,
+        ]);
+
+        if ($response->ok()) {
+            return $response->json('result')[0];
+        } else {
+            self::logError("Registration with id " . $reg_id . " could not be retrieved, reason: " . $response->reason());
             return null;
         }
     }
 
-    public static function getSingleReg(string $id): mixed
+    private static function validateParam(object|string|null $param, string $type): bool
     {
-        $response = Http::get(config('services.regsys.url') . '/dealers-den-api', [
-            'token' => config('services.regsys.token'),
-            'id' => $id,
-        ]);
-
-        if ($response->ok()) {
-            return $response->json();
+        if (empty($param)) {
+            self::logError("Parameter " . $type . " missing");
+            return false;
         } else {
-            Log::warning("Registration with id " . $id . " could not be retrieved, reason: " . $response->reason());
-            return null;
+            return true;
         }
+    }
+
+    private static function logError(string $message)
+    {
+        Log::warning("Class: " . __CLASS__ . ", Function: " . __FUNCTION__ . ": " . $message);
     }
 }
